Seperate queues form graphs and paths
