+++ routes/graph.go
@@ -1,0 +1,57 @@
+package routes
+
+
+// AddNode adds a node to the graph
+func (g *ItemGraph) AddNode(n *Node) {
+	g.Lock.Lock()
+	g.Nodes = append(g.Nodes, n)
+	g.Lock.Unlock()
+}
+
+// AddEdge adds an edge to the graph
+func (g *ItemGraph) AddEdge(n1, n2 *Node, weight int) {
+	g.Lock.Lock()
+	if g.Edges == nil {
+		g.Edges = make(map[Node][]*Edge)
+	}
+	ed1 := Edge{
+		Node:   n2,
+		Weight: weight,
+	}
+
+	ed2 := Edge{
+		Node:   n1,
+		Weight: weight,
+	}
+	g.Edges[*n1] = append(g.Edges[*n1], &ed1)
+	g.Edges[*n2] = append(g.Edges[*n2], &ed2)
+	g.Lock.Unlock()
+}
+
+func CreateInputGraph(inputData *[]InputData, from, to string) InputGraph {
+	inputGraph := InputGraph{
+		Graph: *inputData,
+		From:  from,
+		To:    to,
+	}
+	return inputGraph
+}
+
+func CreateGraph(data InputGraph) *ItemGraph {
+	var g ItemGraph
+	nodes := make(map[string]*Node)
+	for _, v := range data.Graph {
+		if _, found := nodes[v.Source]; !found {
+			nA := Node{v.Source}
+			nodes[v.Source] = &nA
+			g.AddNode(&nA)
+		}
+		if _, found := nodes[v.Destination]; !found {
+			nA := Node{v.Destination}
+			nodes[v.Destination] = &nA
+			g.AddNode(&nA)
+		}
+		g.AddEdge(nodes[v.Source], nodes[v.Destination], v.Weight)
+	}
+	return &g
+}
